move SuggestKetoMealUseCase and KetoFriendlyMealUI into suggest package for better navigation
