+++ src/main/kotlin/di/UseCaseModule.kt
@@ -3,6 +3,7 @@ package org.example.di
 import logic.usecase.*
 import org.example.logic.usecase.suggest.EggFreeSweetsUseCase
 import org.example.logic.usecase.suggest.IraqiMealsIdentifierUseCase
+import org.example.logic.usecase.suggest.SuggestKetoMealUseCase
 import org.koin.dsl.module
 
 val useCaseModule = module {
